+++ miataru/miataru/views/iPad/Devices Views/iPad_DeviceMapView.swift
@@ -992,7 +992,8 @@ struct iPad_DeviceMapView: View {
             let locations = try await MiataruAPIClient.getLocation(
                 serverURL: url,
                 forDeviceIDs: [deviceID],
-                requestingDeviceID: thisDeviceIDManager.shared.deviceID
+                requestingDeviceID: thisDeviceIDManager.shared.deviceID,
+                requestingDeviceKey: settings.deviceKey
             )
             if let loc = locations.first {
                 let coordinate = CLLocationCoordinate2D(latitude: loc.Latitude, longitude: loc.Longitude)