+++ miataru/miataru/views/iPhone/Groups Views/iPhone_GroupMapView.swift
@@ -701,7 +701,8 @@ struct iPhone_GroupMapView: View {
             let locations = try await MiataruAPIClient.getLocation(
                 serverURL: url,
                 forDeviceIDs: groupDeviceIDs,
-                requestingDeviceID: thisDeviceIDManager.shared.deviceID
+                requestingDeviceID: thisDeviceIDManager.shared.deviceID,
+                requestingDeviceKey: settings.deviceKey
             )
             // Update device locations
             for location in locations {
@@ -743,7 +744,8 @@ struct iPhone_GroupMapView: View {
                     let locations = try await MiataruAPIClient.getLocation(
                         serverURL: url,
                         forDeviceIDs: [deviceID],
-                        requestingDeviceID: thisDeviceIDManager.shared.deviceID
+                        requestingDeviceID: thisDeviceIDManager.shared.deviceID,
+                        requestingDeviceKey: settings.deviceKey
                     )
                     if let location = locations.first {
                         let coordinate = CLLocationCoordinate2D(latitude: location.Latitude, longitude: location.Longitude)