+++ miataru/miataruWidgets/SharedWidgetConfig.swift
@@ -13,6 +13,7 @@ struct SharedWidgetConfig: Codable {
     let miataruServerURL: String
     let deviceIDs: [String]
     let ownDeviceID: String?
+    let ownDeviceKey: String?
     let authorizationToken: String?
 }
 
@@ -49,4 +50,3 @@ enum SharedWidgetConfigManager {
         }
     }
 }
-